Remove more legacy code
